requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
motor
python-dotenv

# Зависимости для аутентификации
passlib[bcrypt]
python-jose[cryptography]
python-multipart 